Organize project into src folders
