+++ src/core/game.cpp
@@ -1,16 +1,16 @@
-#include "lcd_display.h"
-#include "game.h"
-#include "game_301.h"
-#include "game_501.h"
-#include "game_shanghai.h"
-#include "game_hangman.h"
-#include "game_roulette.h"
-#include "game_701.h"
-#include "game_cricket.h"
-#include "game_3inline.h"
-#include "config.h"
-#include "melodies.h"
-#include "scoreboard.h"
+#include "modules/lcd_display.h"
+#include "core/game.h"
+#include "games/game_301.h"
+#include "games/game_501.h"
+#include "games/game_shanghai.h"
+#include "games/game_hangman.h"
+#include "games/game_roulette.h"
+#include "games/game_701.h"
+#include "games/game_cricket.h"
+#include "games/game_3inline.h"
+#include "core/config.h"
+#include "modules/melodies.h"
+#include "modules/scoreboard.h"
 
 bool DOUBLE_IN = false;
 bool DOUBLE_OUT = false;
