+++ src/core/game.h
