+++ src/games/game_301.cpp
@@ -1,9 +1,9 @@
-#include "lcd_display.h"
-#include "game_301.h"
-#include "game.h"
-#include "config.h"
-#include "scoreboard.h"
-#include "melodies.h"
+#include "modules/lcd_display.h"
+#include "games/game_301.h"
+#include "core/game.h"
+#include "core/config.h"
+#include "modules/scoreboard.h"
+#include "modules/melodies.h"
 
 int vrijednostMete(const String& naziv) {
     if (naziv.startsWith("Triple ")) return 3 * naziv.substring(7).toInt();
