+++ src/games/game_shanghai.h
